+++ plugin-dev/Source/Sentry/Private/HAL/PlatformSentryHint.h
@@ -8,6 +8,10 @@
 #include "Android/AndroidSentryHint.h"
 #elif PLATFORM_APPLE && !USE_SENTRY_NATIVE
 #include "Apple/AppleSentryHint.h"
+#elif USE_SENTRY_NATIVE && PLATFORM_MICROSOFT
+#include "Microsoft/MicrosoftSentryHint.h"
+#elif USE_SENTRY_NATIVE
+#include "GenericPlatform/GenericPlatformSentryHint.h"
 #else
 #include "Null/NullSentryHint.h"
 #endif